Add an initial /services/ endpoint
This sets up a serializer which can be used in models by inheriting from `Serializer`.

One caveat to this approach is that we need to pull out any relational objects,
because the serialization doesn't handle nested relations. We can either leave this
as is and use RESTful endpoints to get related data, or develop a way to handle
serializing nested data.

You can test this by spinning up the service, running seeds, and hitting:
```
http://localhost:5000/services/
```

You should get a response back similar to:
```
[
  {
    "branch": "master",
    "deploy_file": null,
    "display_name": "RBAC",
    "gh_repo": "https://github.com/RedHatInsights/insights-rbac/",
    "gl_repo": null,
    "id": 1,
    "name": "rbac",
    "namespace": null
  },
  {
    "branch": "master",
    "deploy_file": null,
    "display_name": "Entitlements",
    "gh_repo": "https://github.com/RedHatInsights/entitlements-api-go/",
    "gl_repo": null,
    "id": 2,
    "name": "entitlements",
    "namespace": null
  }
]
```

The next step will be to have the UI render a table with this data.