[refactor] split out dev-db seeding from prod db
The prod db needs to use the seeds() function to get the services loaded
into the database. In dev, we want to have the services AND some commits
so we can test sorting and such. This creates a new flask command that
will take care of that.

`FLASK_APP=api flask dev-db`

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>